Simplify getting next and prev Herbaria
- Make them part of `herbaria#show` instead of separate, namespaced actions
- Delete old namespaced HerbariaNextsController, including tests
- update routes
- update links to `next` and `prev`
